Crowler/pyproject.toml
[project]
name = "jobvite-crowler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "selenium>=4.35.0",
    "python-dotenv>=1.0.0",
]
